+++ meridian/analysis/review/reviewer.py
@@ -26,9 +26,11 @@ from meridian.analysis import analyzer as analyzer_module
 from meridian.analysis.review import checks
 from meridian.analysis.review import configs
 from meridian.analysis.review import constants as review_constants
+
 from meridian.analysis.review import results
 from meridian.model import context
 from meridian.model import prior_distribution
+
 import numpy as np
 
 CheckType = typing.Type[checks.BaseCheck]